refactor(core): extract step callbacks from FrontAgent
Move executeSteps callback logic (onStepComplete, onPhaseError,
onPhaseComplete) into step-callbacks.ts with factory functions,
reducing agent.ts from 1334 to 1028 lines.

Closes #114
